Removed Impl. Folder from Repos, change to file scope namespaces
