+++ Project.App/Project.Api/Data/MappingProfile.cs
@@ -3,10 +3,8 @@ using AutoMapper;
 using Project.Api.Models;
 using Project.DTOs;
 
-namespace Project.Api.Data
+namespace Project.Api.Data;
+public class MappingProfile : Profile
 {
-    public class MappingProfile : Profile
-    {
-        public MappingProfile() { }
-    }
+    public MappingProfile() { }
 }
