+++ Project.App/Project.Api/Models/Room.cs
@@ -1,8 +1,8 @@
-namespace Project.Api.Models;
 
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
+namespace Project.Api.Models;
 public class Room
 {
     [Key]
