+++ Project.App/Project.Api/Repositories/Interface/IUserRepository.cs
@@ -1,15 +1,15 @@
 using Project.Api.Models;
 
-namespace Project.Api.Repositories.Interface
+namespace Project.Api.Repositories.Interface;
+
+public interface IUserRepository
 {
-    public interface IUserRepository
-    {
-        Task<IEnumerable<User>> GetAllAsync();
-        Task<User?> GetByIdAsync(int id);
-        Task<User?> GetByEmailAsync(string email);
-        Task AddAsync(User user);
-        Task UpdateAsync(User user);
-        Task DeleteAsync(int id);
-        Task SaveChangesAsync();
-    }
+    Task<IEnumerable<User>> GetAllAsync();
+    Task<User?> GetByIdAsync(int id);
+    Task<User?> GetByEmailAsync(string email);
+    Task AddAsync(User user);
+    Task UpdateAsync(User user);
+    Task DeleteAsync(int id);
+    Task SaveChangesAsync();
 }
+
