+++ Project.App/Project.Api/Services/HandService.cs
@@ -16,131 +16,129 @@ using Project.Api.Models;
 using Project.Api.Repositories;
 using Project.Api.Services;
 
-namespace Project.Api.Services
+namespace Project.Api.Services;
+public class HandService : IHandService
 {
-    public class HandService : IHandService
+    // Call the repository
+    private readonly IHandRepository _Repo;
+    private readonly ILogger<HandService> _logger;
+
+    // Constructor
+    public HandService(IHandRepository repo, ILogger<HandService> logger)
     {
-        // Call the repository
-        private readonly IHandRepository _Repo;
-        private readonly ILogger<HandService> _logger;
+        // Dependency Injection
+        _Repo = repo;
+        _logger = logger;
+    }
 
-        // Constructor
-        public HandService(IHandRepository repo, ILogger<HandService> logger)
-        {
-            // Dependency Injection
-            _Repo = repo;
-            _logger = logger;
-        }
-
-        // Implement the methods
-        // Create a new hand
-        public async Task<Hand> CreateHandAsync(Hand hand)
+    // Implement the methods
+    // Create a new hand
+    public async Task<Hand> CreateHandAsync(Hand hand)
+    {
+        // send request to repository
+        _logger.LogInformation("Creating a new hand");
+        return await _Repo.CreateHandAsync(hand);
+    }
+    // Get all hands in a room
+    public async Task<List<Hand>> GetHandsByRoomIdAsync(Guid roomId)
+    {
+        try
         {
             // send request to repository
-            _logger.LogInformation("Creating a new hand");
-            return await _Repo.CreateHandAsync(hand);
+            _logger.LogInformation($"Getting all hands for room {roomId}");
+            return await _Repo.GetHandsByRoomIdAsync(roomId);
         }
-        // Get all hands in a room
-        public async Task<List<Hand>> GetHandsByRoomIdAsync(Guid roomId)
+        catch (Exception e)
         {
-            try
-            {
-                // send request to repository
-                _logger.LogInformation($"Getting all hands for room {roomId}");
-                return await _Repo.GetHandsByRoomIdAsync(roomId);
-            }
-            catch (Exception e)
-            {
-                // throw an exception and Log it
-                _logger.LogError(e, $"Error getting hands for room {roomId}: {e.Message}");
-                throw new Exception(e.Message);
-            }
+            // throw an exception and Log it
+            _logger.LogError(e, $"Error getting hands for room {roomId}: {e.Message}");
+            throw new Exception(e.Message);
         }
-        // Get a hand by its ID
-        public async Task<Hand?> GetHandAsyncByIdAsync(Guid handId)
+    }
+    // Get a hand by its ID
+    public async Task<Hand?> GetHandAsyncByIdAsync(Guid handId)
+    {
+        try
         {
-            try
-            {
-                // send request to repository
-                _logger.LogInformation($"Getting hand {handId}");
-                return await _Repo.GetHandAsyncById(handId);
-            }
-            catch (Exception e)
-            {
-                // throw an exception and Log it
-                _logger.LogError(e, $"Error getting hand {handId}: {e.Message}");
-                throw new Exception(e.Message);
-            }
+            // send request to repository
+            _logger.LogInformation($"Getting hand {handId}");
+            return await _Repo.GetHandAsyncById(handId);
         }
-        // Get all hands by a user in a room
-        public async Task<List<Hand>> GetHandsByUserIdAsync(Guid roomId, Guid userId)
+        catch (Exception e)
         {
-            try
-            {
-                // send request to repository
-                _logger.LogInformation($"Getting all hands for user {userId} in room {roomId}");
-                return await _Repo.GetHandsByUserIdAsync(roomId, userId);
-            }
-            catch (Exception e)
-            {
-                // throw an exception and Log it
-                _logger.LogError(e, $"Error getting all hands for user {userId} in room {roomId}: {e.Message}");
-                throw new Exception(e.Message);
-            }
-            
+            // throw an exception and Log it
+            _logger.LogError(e, $"Error getting hand {handId}: {e.Message}");
+            throw new Exception(e.Message);
         }
-
-
-        // Update an existing hand
-        public async Task<Hand> UpdateHandAsync(Guid handId, Hand hand)
+    }
+    // Get all hands by a user in a room
+    public async Task<List<Hand>> GetHandsByUserIdAsync(Guid roomId, Guid userId)
+    {
+        try
         {
-            try
-            {
-                // send request to repository
-                _logger.LogInformation($"Updating hand {handId}");
-                return await _Repo.UpdateHandAsync(handId, hand);
-            }
-            catch (Exception e)
-            {
-                // throw an exception and Log it
-                _logger.LogError(e, $"Error updating hand {handId}: {e.Message}");
-                throw new Exception(e.Message);
-            }
+            // send request to repository
+            _logger.LogInformation($"Getting all hands for user {userId} in room {roomId}");
+            return await _Repo.GetHandsByUserIdAsync(roomId, userId);
         }
-        // Partially update an existing hand
-        public async Task<Hand> PatchHandAsync(Guid handId, int? Order = null, string? CardsJson = null, int? Bet = null)
+        catch (Exception e)
         {
-            try
-            {
-                // send request to repository
-                _logger.LogInformation($"Patching hand {handId}");
-                return await _Repo.PatchHandAsync(handId, Order, CardsJson, Bet);
-            }
-            catch (Exception e)
-            {
-                // throw an exception and Log it
-                _logger.LogError(e, $"Error patching hand {handId}: {e.Message}");
-                throw new Exception(e.Message);
-            }
-        }
-        // Delete a hand
-        public async Task<Hand> DeleteHandAsync(Guid handId)
-        {
-            try
-            {
-                // send request to repository
-                _logger.LogInformation($"Deleting hand {handId}");
-                return await _Repo.DeleteHandAsync(handId);
-            }
-            catch (Exception e)
-            {
-                // throw an exception and Log it
-                _logger.LogError(e, $"Error deleting hand {handId}: {e.Message}");
-                throw new Exception(e.Message);
-            }
-            
+            // throw an exception and Log it
+            _logger.LogError(e, $"Error getting all hands for user {userId} in room {roomId}: {e.Message}");
+            throw new Exception(e.Message);
         }
         
-
     }
+
+
+    // Update an existing hand
+    public async Task<Hand> UpdateHandAsync(Guid handId, Hand hand)
+    {
+        try
+        {
+            // send request to repository
+            _logger.LogInformation($"Updating hand {handId}");
+            return await _Repo.UpdateHandAsync(handId, hand);
+        }
+        catch (Exception e)
+        {
+            // throw an exception and Log it
+            _logger.LogError(e, $"Error updating hand {handId}: {e.Message}");
+            throw new Exception(e.Message);
+        }
+    }
+    // Partially update an existing hand
+    public async Task<Hand> PatchHandAsync(Guid handId, int? Order = null, string? CardsJson = null, int? Bet = null)
+    {
+        try
+        {
+            // send request to repository
+            _logger.LogInformation($"Patching hand {handId}");
+            return await _Repo.PatchHandAsync(handId, Order, CardsJson, Bet);
+        }
+        catch (Exception e)
+        {
+            // throw an exception and Log it
+            _logger.LogError(e, $"Error patching hand {handId}: {e.Message}");
+            throw new Exception(e.Message);
+        }
+    }
+    // Delete a hand
+    public async Task<Hand> DeleteHandAsync(Guid handId)
+    {
+        try
+        {
+            // send request to repository
+            _logger.LogInformation($"Deleting hand {handId}");
+            return await _Repo.DeleteHandAsync(handId);
+        }
+        catch (Exception e)
+        {
+            // throw an exception and Log it
+            _logger.LogError(e, $"Error deleting hand {handId}: {e.Message}");
+            throw new Exception(e.Message);
+        }
+        
+    }
+    
+
 }
