+++ Project.App/Project.Api/Services/Interface/IHandService.cs
@@ -5,23 +5,21 @@
 */
 using Project.Api.Models;
 
-namespace Project.Api.Services
+namespace Project.Api.Services;
+public interface IHandService
 {
-    public interface IHandService
-    {
-        // Define Fields
-        Task<List<Hand>> GetHandsByRoomIdAsync(Guid roomId);
+    // Define Fields
+    Task<List<Hand>> GetHandsByRoomIdAsync(Guid roomId);
 
-        Task<Hand?> GetHandAsyncByIdAsync(Guid handId);
+    Task<Hand?> GetHandAsyncByIdAsync(Guid handId);
 
-        Task<List<Hand>> GetHandsByUserIdAsync(Guid roomId, Guid userId);
+    Task<List<Hand>> GetHandsByUserIdAsync(Guid roomId, Guid userId);
 
-        Task<Hand> CreateHandAsync(Hand hand);
+    Task<Hand> CreateHandAsync(Hand hand);
 
-        Task<Hand> UpdateHandAsync(Guid handId, Hand hand);
+    Task<Hand> UpdateHandAsync(Guid handId, Hand hand);
 
-        Task<Hand> PatchHandAsync(Guid handId, int? Order = null, string? CardsJson = null, int? Bet = null);
+    Task<Hand> PatchHandAsync(Guid handId, int? Order = null, string? CardsJson = null, int? Bet = null);
 
-        Task<Hand> DeleteHandAsync(Guid handId);
-    }
+    Task<Hand> DeleteHandAsync(Guid handId);
 }
