--- Project.App/Project.Api/Services/Interface/IServiceExample.cs
@@ -1,6 +1,0 @@
-using Project.Api.Models;
-
-namespace Project.Api.Services
-{
-    public interface IServiceExample { }
-}
